changed long long to int64_t

+++ src/ConflictTable.h
@@ -2,19 +2,20 @@
 #define __CONFLICTTABLE_H__
 
 #include <list>
+#include "Types.h"
 
 using namespace std;
 
 public
 ConflictTable {
   public:
-    ConflictTable(long long startPC, long long endPC){};
+    ConflictTable(int64_t startPC, int64_t endPC){};
 
     int getNumConflicts() { return myDependences.size(); };
 
   private:
-    long long myLoopStartPC;
-    long long myLoopEndPC;
+    int64_t myLoopStartPC;
+    int64_t myLoopEndPC;
 
     list<Dependence> myDependences;
 };
