+++ src/Dependence.h
@@ -1,5 +1,6 @@
 #include __DEPENDENCE_H__
 #define __DEPENDENCE_H__
+#include "Types.h"
 
 enum DependenceType { NODEP = -1, RAW = 0, WAR = 1, WAW = 2 };
 
@@ -7,21 +8,21 @@ class Dependence {
   public:
     Dependence();
 
-    Dependence(long long memoryAddr, long long earlierPC, MemAccessMode earlierAccessMode, long long laterPC,
+    Dependence(int64_t memoryAddr, int64_t earlierPC, MemAccessMode earlierAccessMode, int64_t laterPC,
                MemAccessMode laterAccessMode);
 
     DependenceType getDependenceType() { return myDependenceType; };
 
-    long long getMemoryAddress() { return myMemoryAddr; };
+    int64_t getMemoryAddress() { return myMemoryAddr; };
 
   private:
     DependenceType myDependenceType;
-    long long myMemoryAddr;
+    int64_t myMemoryAddr;
 
-    long long myEarlierPC;
+    int64_t myEarlierPC;
     MemAccessMode myEarlierAccessMode;
 
-    long long myLaterPC;
+    int64_t myLaterPC;
     MemAccessMode myLaterAccessMode;
 };
