+++ src/PendingStrideTable.cpp
@@ -1,27 +1,27 @@
 #include "PendingStrideTable.h"
 
 // This method assumes that a stride does not already exist
-void PendingStrideTable::addNewStride(long long PC, Stride *stride) {
+void PendingStrideTable::addNewStride(int64_t PC, Stride *stride) {
     StrideEntry temp;
     temp.stride = stride;
     temp.killed = false;
 
-    myPendingStrides.insert(pair<long long, StrideEntry>(PC, temp));
+    myPendingStrides.insert(pair<int64_t, StrideEntry>(PC, temp));
 }
 
-void PendingStrideTable::addNewStride(long long PC, long long lowAddress, long long highAddress, long long strideLength,
-                                      int numAccesses, long long accessSize, MemAccessMode readOrWrite) {
+void PendingStrideTable::addNewStride(int64_t PC, int64_t lowAddress, int64_t highAddress, int64_t strideLength,
+                                      int numAccesses, int64_t accessSize, MemAccessMode readOrWrite) {
     StrideEntry temp;
     temp.stride = new Stride(lowAddress, highAddress, strideLength, numAccesses, accessSize, readOrWrite);
     temp.killed = false;
 
-    myPendingStrides.insert(pair<long long, StrideEntry>(PC, temp));
+    myPendingStrides.insert(pair<int64_t, StrideEntry>(PC, temp));
 }
 
-bool PendingStrideTable::doesStrideExist(long long PC, long long memoryAddr, long long strideLength,
+bool PendingStrideTable::doesStrideExist(int64_t PC, int64_t memoryAddr, int64_t strideLength,
                                          MemAccessMode readOrWrite) {
-    pair<multimap<long long, StrideEntry>::iterator, multimap<long long, StrideEntry>::iterator> rangeIter;
-    multimap<long long, StrideEntry>::iterator iter;
+    pair<multimap<int64_t, StrideEntry>::iterator, multimap<int64_t, StrideEntry>::iterator> rangeIter;
+    multimap<int64_t, StrideEntry>::iterator iter;
 
     rangeIter = myPendingStrides.equal_range(PC);
 
@@ -47,10 +47,10 @@ bool PendingStrideTable::doesStrideExist(long long PC, long long memoryAddr, lon
     return false;
 }
 
-bool PendingStrideTable::updateExistingStride(long long PC, long long memoryAddr, long long strideLength,
+bool PendingStrideTable::updateExistingStride(int64_t PC, int64_t memoryAddr, int64_t strideLength,
                                               MemAccessMode readOrWrite) {
-    pair<multimap<long long, StrideEntry>::iterator, multimap<long long, StrideEntry>::iterator> rangeIter;
-    multimap<long long, StrideEntry>::iterator iter;
+    pair<multimap<int64_t, StrideEntry>::iterator, multimap<int64_t, StrideEntry>::iterator> rangeIter;
+    multimap<int64_t, StrideEntry>::iterator iter;
 
     rangeIter = myPendingStrides.equal_range(PC);
 
@@ -65,10 +65,10 @@ bool PendingStrideTable::updateExistingStride(long long PC, long long memoryAddr
     return false;
 }
 
-bool PendingStrideTable::killStride(long long PC, long long memoryAddr, long long strideLength,
+bool PendingStrideTable::killStride(int64_t PC, int64_t memoryAddr, int64_t strideLength,
                                     MemAccessMode readOrWrite) {
-    pair<multimap<long long, StrideEntry>::iterator, multimap<long long, StrideEntry>::iterator> rangeIter;
-    multimap<long long, StrideEntry>::iterator iter;
+    pair<multimap<int64_t, StrideEntry>::iterator, multimap<int64_t, StrideEntry>::iterator> rangeIter;
+    multimap<int64_t, StrideEntry>::iterator iter;
 
     rangeIter = myPendingStrides.equal_range(PC);
 
@@ -86,7 +86,7 @@ bool PendingStrideTable::killStride(long long PC, long long memoryAddr, long lon
 }
 
 void PendingStrideTable::print() {
-    multimap<long long, StrideEntry>::iterator iter;
+    multimap<int64_t, StrideEntry>::iterator iter;
 
     for (iter = myPendingStrides.begin(); iter != myPendingStrides.end(); iter++) {
         cout << "PC: " << iter->first;
