+++ src/Point.h
@@ -11,21 +11,21 @@ class Point {
   public:
     // Create a new point including its memory address and the PC that made the
     // memory access
-    Point(long long memAddress, long long PC, long long numAccesses = 1, long long accessSize = 4,
+    Point(int64_t memAddress, int64_t PC, int64_t numAccesses = 1, int64_t accessSize = 4,
           MemAccessMode readOrWrite = READ);
 
     // Return the PC associated with this memory access
-    long long getPC() { return myPC; };
+    int64_t getPC() { return myPC; };
 
     // Return this point's memory address
-    long long getMemAddress() { return myMemAddress; };
+    int64_t getMemAddress() { return myMemAddress; };
 
     // Return the number of times this memory address has been accessed by this
     // PC
-    long long getNumAccesses() { return myNumAccesses; };
+    int64_t getNumAccesses() { return myNumAccesses; };
 
     // Return the size of the memory access
-    long long getAccessSize() { return myAccessSize; };
+    int64_t getAccessSize() { return myAccessSize; };
 
     // Return the access mode (read or  write)
     MemAccessMode getAccessMode() { return myAccessMode; };
@@ -36,10 +36,10 @@ class Point {
     void print();
 
   private:
-    long long myPC;
-    long long myMemAddress;
-    long long myNumAccesses;
-    long long myAccessSize;
+    int64_t myPC;
+    int64_t myMemAddress;
+    int64_t myNumAccesses;
+    int64_t myAccessSize;
     MemAccessMode myAccessMode; // read or write
 };
